+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Implementation changes and bug fixes
 - [PR #2029](https://github.com/rqlite/rqlite/pull/2029): Add basic preupdate hook support at database level.
 - [PR #2030](https://github.com/rqlite/rqlite/pull/2030), [PR #2031](https://github.com/rqlite/rqlite/pull/2031): Improve preupdate hook stats and error handling.
+- [PR #2032](https://github.com/rqlite/rqlite/pull/2032): Support preupdate hook for row ID values only.
 
 ## v8.36.6 (January 19th 2025)
 ### Implementation changes and bug fixes